src/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==4.2.7
django-ninja==1.2.0
gunicorn==22.0.0
packaging==24.1
pillow==10.4.0
pydantic==2.7.4
pydantic_core==2.18.4
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
gunicorn
